# isOpaque, faster nonsimd draw, drawBasic