--- a/src/pixie.nim
+++ b/src/pixie.nim
@@ -1,10 +1,9 @@
-import bumpy, chroma, flatty/binny, os, pixie/blends, pixie/common,
-    pixie/contexts, pixie/fileformats/bmp, pixie/fileformats/gif,
-    pixie/fileformats/jpg, pixie/fileformats/png, pixie/fileformats/svg,
-    pixie/fonts, pixie/images, pixie/masks, pixie/paints, pixie/paths, strutils, vmath
+import bumpy, chroma, flatty/binny, os, pixie/common, pixie/contexts,
+    pixie/fileformats/bmp, pixie/fileformats/gif, pixie/fileformats/jpg,
+    pixie/fileformats/png, pixie/fileformats/svg, pixie/fonts, pixie/images,
+    pixie/masks, pixie/paints, pixie/paths, strutils, vmath
 
-export blends, bumpy, chroma, common, contexts, fonts, images, masks, paints,
-    paths, vmath
+export bumpy, chroma, common, contexts, fonts, images, masks, paints, paths, vmath
 
 type
   FileFormat* = enum