--- a/src/pixie/images.nim
+++ b/src/pixie/images.nim
@@ -913,29 +913,70 @@ proc drawUber(
         clamp(srcPos.y, 0, b.height.float32)
       )
 
-      for x in x ..< xMax:
-        let samplePos = ivec2((srcPos.x - h).int32, (srcPos.y - h).int32)
-
-        when type(a) is Image:
-          let backdrop = a.unsafe[x, y]
-          when type(b) is Image:
-            let
-              sample = b.unsafe[samplePos.x, samplePos.y]
-              blended = blender(backdrop, sample)
-          else: # b is a Mask
-            let
-              sample = b.unsafe[samplePos.x, samplePos.y]
-              blended = blender(backdrop, rgbx(0, 0, 0, sample))
-          a.unsafe[x, y] = blended
-        else: # a is a Mask
-          let backdrop = a.unsafe[x, y]
-          when type(b) is Image:
-            let sample = b.unsafe[samplePos.x, samplePos.y].a
-          else: # b is a Mask
-            let sample = b.unsafe[samplePos.x, samplePos.y]
-          a.unsafe[x, y] = masker(backdrop, sample)
-
-        srcPos += dx
+      if blendMode == bmOverwrite:
+        for x in x ..< xMax:
+          let samplePos = ivec2((srcPos.x - h).int32, (srcPos.y - h).int32)
+          when type(a) is Image:
+            when type(b) is Image:
+              let source = b.unsafe[samplePos.x, samplePos.y]
+            else: # b is a Mask
+              let source = rgbx(0, 0, 0, b.unsafe[samplePos.x, samplePos.y])
+            a.unsafe[x, y] = source
+          else: # a is a Mask
+            when type(b) is Image:
+              let source = b.unsafe[samplePos.x, samplePos.y].a
+            else: # b is a Mask
+              let source = b.unsafe[samplePos.x, samplePos.y]
+            a.unsafe[x, y] = source
+          srcPos += dx
+      elif blendMode == bmNormal:
+        for x in x ..< xMax:
+          let samplePos = ivec2((srcPos.x - h).int32, (srcPos.y - h).int32)
+          when type(a) is Image:
+            when type(b) is Image:
+              let source = b.unsafe[samplePos.x, samplePos.y]
+            else: # b is a Mask
+              let source = rgbx(0, 0, 0, b.unsafe[samplePos.x, samplePos.y])
+            if source.a > 0:
+              if source.a == 255:
+                a.unsafe[x, y] = source
+              else:
+                let backdrop = a.unsafe[x, y]
+                a.unsafe[x, y] = blendNormal(backdrop, source)
+          else: # a is a Mask
+            when type(b) is Image:
+              let source = b.unsafe[samplePos.x, samplePos.y].a
+            else: # b is a Mask
+              let source = b.unsafe[samplePos.x, samplePos.y]
+            if source > 0:
+              if source == 255:
+                a.unsafe[x, y] = source
+              else:
+                let backdrop = a.unsafe[x, y]
+                a.unsafe[x, y] = blendAlpha(backdrop, source)
+          srcPos += dx
+      else:
+        for x in x ..< xMax:
+          let samplePos = ivec2((srcPos.x - h).int32, (srcPos.y - h).int32)
+          when type(a) is Image:
+            let backdrop = a.unsafe[x, y]
+            when type(b) is Image:
+              let
+                sample = b.unsafe[samplePos.x, samplePos.y]
+                blended = blender(backdrop, sample)
+            else: # b is a Mask
+              let
+                sample = b.unsafe[samplePos.x, samplePos.y]
+                blended = blender(backdrop, rgbx(0, 0, 0, sample))
+            a.unsafe[x, y] = blended
+          else: # a is a Mask
+            let backdrop = a.unsafe[x, y]
+            when type(b) is Image:
+              let sample = b.unsafe[samplePos.x, samplePos.y].a
+            else: # b is a Mask
+              let sample = b.unsafe[samplePos.x, samplePos.y]
+            a.unsafe[x, y] = masker(backdrop, sample)
+          srcPos += dx
 
     if blendMode == bmMask:
       if a.width - xMax > 0: