`nim c -d:release -d:pixieNoSimd --gc:arc --debugger:native -r .\experiments\benchmark_cairo_draw.nim`

before:
```
pixie draw normal .................. 6.336 ms      6.629 ms    ±0.076   x751
pixie draw overwrite ............... 5.663 ms      5.949 ms    ±0.174   x836
pixie draw mask .................... 9.015 ms      9.099 ms    ±0.097   x547
```

after:
```
pixie draw normal .................. 2.663 ms      2.676 ms    ±0.022  x1000
pixie draw overwrite ............... 3.203 ms      3.266 ms    ±0.096  x1000
pixie draw mask .................... 2.685 ms      2.711 ms    ±0.030  x1000
```

```
pixie draw basic ................... 0.411 ms      0.423 ms    ±0.016  x1000
```